Add `SkipForTigris` helper, use it
