+++ integration/delete_test.go
@@ -27,6 +27,8 @@ import (
 
 // TestDeleteSimple checks simple cases of doc deletion.
 func TestDeleteSimple(t *testing.T) {
+	setup.SkipForTigris(t)
+
 	t.Parallel()
 	ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
