+++ integration/setup/setup.go
@@ -12,43 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package setup provides integration tests setup helpers.
 package setup
 
 import (
 	"context"
-	"flag"
-	"fmt"
-	"net"
-	"net/url"
-	"sync"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zaptest"
 
 	"github.com/FerretDB/FerretDB/integration/shareddata"
-	"github.com/FerretDB/FerretDB/internal/clientconn"
-	"github.com/FerretDB/FerretDB/internal/handlers/registry"
-	"github.com/FerretDB/FerretDB/internal/util/debug"
-	"github.com/FerretDB/FerretDB/internal/util/logging"
 	"github.com/FerretDB/FerretDB/internal/util/testutil"
-)
-
-var (
-	targetPortF = flag.Int("target-port", 0, "target system's port for tests; if 0, in-process FerretDB is used")
-	proxyAddrF  = flag.String("proxy-addr", "", "proxy to use for in-process FerretDB")
-	handlerF    = flag.String("handler", "pg", "handler to use for in-process FerretDB")
-
-	// Disable noisy setup logs by default.
-	debugSetupF = flag.Bool("debug-setup", false, "enable debug logs for tests setup")
-	logLevelF   = zap.LevelFlag("log-level", zap.DebugLevel, "log level for tests")
-
-	startupOnce sync.Once
 )
 
 // SetupOpts represents setup options.
@@ -73,11 +49,7 @@ type SetupResult struct {
 func SetupWithOpts(tb testing.TB, opts *SetupOpts) *SetupResult {
 	tb.Helper()
 
-	startupOnce.Do(func() {
-		logging.Setup(zap.DebugLevel)
-
-		go debug.RunHandler(testutil.Ctx(tb), "127.0.0.1:0", zap.L().Named("debug"))
-	})
+	startup()
 
 	if opts == nil {
 		opts = new(SetupOpts)
@@ -99,7 +71,13 @@ func SetupWithOpts(tb testing.TB, opts *SetupOpts) *SetupResult {
 	// register cleanup function after setupListener registers its own to preserve full logs
 	tb.Cleanup(cancel)
 
-	collection := setupCollection(tb, ctx, port, opts.DatabaseName, opts.Providers)
+	client := setupClient(tb, ctx, port)
+
+	collection := setupCollection(tb, ctx, &setupCollectionOpts{
+		client:    client,
+		db:        opts.DatabaseName,
+		providers: opts.Providers,
+	})
 
 	level.SetLevel(*logLevelF)
 
@@ -120,74 +98,18 @@ func Setup(tb testing.TB, providers ...shareddata.Provider) (context.Context, *m
 	return s.Ctx, s.Collection
 }
 
-// setupListener starts in-process FerretDB server that runs until ctx is done,
-// and returns listening port number.
-func setupListener(tb testing.TB, ctx context.Context, logger *zap.Logger) int {
+// setupCollection setups a single collection for all providers, if the are present.
+func setupCollection(tb testing.TB, ctx context.Context, opts *setupCollectionOpts) *mongo.Collection {
 	tb.Helper()
-
-	h, err := registry.NewHandler(*handlerF, &registry.NewHandlerOpts{
-		Ctx:           ctx,
-		Logger:        logger,
-		PostgreSQLURL: testutil.PoolConnString(tb, nil),
-		TigrisURL:     testutil.TigrisURL(tb),
-	})
-	require.NoError(tb, err)
-
-	proxyAddr := *proxyAddrF
-	mode := clientconn.NormalMode
-	if proxyAddr != "" {
-		mode = clientconn.DiffNormalMode
-	}
-
-	l := clientconn.NewListener(&clientconn.NewListenerOpts{
-		ListenAddr:         "127.0.0.1:0",
-		ProxyAddr:          proxyAddr,
-		Mode:               mode,
-		Handler:            h,
-		Logger:             logger,
-		TestRunCancelDelay: time.Hour, // make it easier to notice missing client's disconnects
-	})
-
-	done := make(chan struct{})
-	go func() {
-		defer close(done)
-
-		err := l.Run(ctx)
-		if err == nil || err == context.Canceled {
-			logger.Info("Listener stopped")
-		} else {
-			logger.Error("Listener stopped", zap.Error(err))
-		}
-	}()
-
-	// ensure that all listener's logs are written before test ends
-	tb.Cleanup(func() {
-		<-done
-		h.Close()
-	})
-
-	return l.Addr().(*net.TCPAddr).Port
-}
-
-// setupCollection setups a single collection.
-// If there are no providers, we don't create a database and collection.
-// That is intentional:
-//   * for those tests where no collection and database are needed.
-//   * for Tigris: we can't create a collection without a schema, and we don't know schema without documents.
-func setupCollection(tb testing.TB, ctx context.Context, port int, db string, providers []shareddata.Provider) *mongo.Collection {
-	tb.Helper()
-
-	require.Greater(tb, port, 0)
-	require.Less(tb, port, 65536)
 
 	var ownDatabase bool
+	db := opts.db
 	if db == "" {
 		db = testutil.DatabaseName(tb)
 		ownDatabase = true
 	}
 
-	client := setupClient(tb, ctx, uint16(port))
-	database := client.Database(db)
+	database := opts.client.Database(db)
 	collectionName := testutil.CollectionName(tb)
 	collection := database.Collection(collectionName)
 
@@ -197,7 +119,7 @@ func setupCollection(tb testing.TB, ctx context.Context, port int, db string, pr
 		_ = database.Drop(ctx)
 	}
 
-	for _, provider := range providers {
+	for _, provider := range opts.providers {
 		docs := shareddata.Docs(provider)
 		require.NotEmpty(tb, docs)
 
@@ -209,7 +131,7 @@ func setupCollection(tb testing.TB, ctx context.Context, port int, db string, pr
 	// delete collection and (possibly) database unless test failed
 	tb.Cleanup(func() {
 		if tb.Failed() {
-			tb.Logf("Keeping database %q and collection %q for debugging.", db, collectionName)
+			tb.Logf("Keeping %s.%s for debugging.", db, collectionName)
 			return
 		}
 
@@ -223,47 +145,4 @@ func setupCollection(tb testing.TB, ctx context.Context, port int, db string, pr
 	})
 
 	return collection
-}
-
-// setupClient returns MongoDB client for database on 127.0.0.1:port.
-func setupClient(tb testing.TB, ctx context.Context, port uint16) *mongo.Client {
-	tb.Helper()
-
-	// those options should not affect anything except tests speed
-	v := url.Values{
-		// TODO: Test fails occured on some platforms due to i/o timeout.
-		// Needs more investigation.
-		//
-		//"connectTimeoutMS":         []string{"5000"},
-		//"serverSelectionTimeoutMS": []string{"5000"},
-		//"socketTimeoutMS":          []string{"5000"},
-		//"heartbeatFrequencyMS":     []string{"30000"},
-
-		//"minPoolSize":   []string{"1"},
-		//"maxPoolSize":   []string{"1"},
-		//"maxConnecting": []string{"1"},
-		//"maxIdleTimeMS": []string{"0"},
-
-		//"directConnection": []string{"true"},
-		//"appName":          []string{tb.Name()},
-	}
-
-	u := url.URL{
-		Scheme:   "mongodb",
-		Host:     fmt.Sprintf("127.0.0.1:%d", port),
-		Path:     "/",
-		RawQuery: v.Encode(),
-	}
-	client, err := mongo.Connect(ctx, options.Client().ApplyURI(u.String()))
-	require.NoError(tb, err)
-
-	err = client.Ping(ctx, nil)
-	require.NoError(tb, err)
-
-	tb.Cleanup(func() {
-		err = client.Disconnect(ctx)
-		require.NoError(tb, err)
-	})
-
-	return client
 }
