+++ integration/update_field_test.go
@@ -33,6 +33,8 @@ import (
 )
 
 func TestUpdateFieldCurrentDate(t *testing.T) {
+	setup.SkipForTigris(t)
+
 	t.Parallel()
 
 	t.Run("readBack", func(t *testing.T) {
@@ -277,6 +279,8 @@ func TestUpdateFieldCurrentDate(t *testing.T) {
 }
 
 func TestUpdateFieldInc(t *testing.T) {
+	setup.SkipForTigris(t)
+
 	t.Parallel()
 
 	t.Run("Ok", func(t *testing.T) {
@@ -628,6 +632,8 @@ func TestUpdateFieldInc(t *testing.T) {
 }
 
 func TestUpdateFieldSet(t *testing.T) {
+	setup.SkipForTigris(t)
+
 	t.Parallel()
 
 	for name, tc := range map[string]struct {
@@ -833,6 +839,8 @@ func TestUpdateFieldSet(t *testing.T) {
 }
 
 func TestUpdateFieldSetOnInsert(t *testing.T) {
+	setup.SkipForTigris(t)
+
 	t.Parallel()
 
 	for name, tc := range map[string]struct {
@@ -944,6 +952,8 @@ func TestUpdateFieldSetOnInsert(t *testing.T) {
 }
 
 func TestUpdateFieldUnset(t *testing.T) {
+	setup.SkipForTigris(t)
+
 	t.Parallel()
 
 	for name, tc := range map[string]struct {
@@ -1023,6 +1033,8 @@ func TestUpdateFieldUnset(t *testing.T) {
 }
 
 func TestUpdateFieldMixed(t *testing.T) {
+	setup.SkipForTigris(t)
+
 	t.Parallel()
 
 	for name, tc := range map[string]struct {
